# server/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "World of Stories API",
    "version": "1.0.0",
    "description": "Documentación limpia y modularizada"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "paths": {
...
    "/npcs/{id}": {
      "get": {
        "operationId": "getNpc",
        "tags": [
          "NPCs"
        ],
        "summary": "Obtener un NPC por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del NPC devuelto con éxito"
          },
          "404": {
            "description": "NPC no encontrado o no publicado."
          }
        }
      },
      "put": {
        "operationId": "updateNpc",
        "tags": [
          "NPCs"
        ],
        "summary": "Actualizar propiedades de un NPC",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "npc": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "items": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "NPC modificado de forma exitosa"
            },
            "404": {
              "description": "No se encontró el NPC para actualizar."
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteNpc",
        "tags": [
          "NPCs"
        ],
        "summary": "Eliminar un NPC de forma global",
        "description": "Borra el NPC y remueve automáticamente su ID de cualquier lista de eventos asociados en ciudades.",
        "parameters": [
          {
            "name": "npcId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "NPC eliminado y desvinculado de su lugar de origen correctamente."
          },
          "404": {
            "description": "El NPC que intentas eliminar no existe."
          }
        }
      }
    },
    "/npcs/{id}/publish": {
      "patch": {
        "operationId": "publishNpc",
        "tags": [
          "NPCs"
        ],
        "summary": "Publicar o despublicar un NPC de forma parcial",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "published": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado de publicación actualizado con éxito."
          }
        }
      }
    },
...
  },
...
}